move header logic
