+++ src/app/app.component.html
@@ -1,43 +1,4 @@
-<mat-toolbar color="primary" class="toolbar">
-  <span>CV Builder<sub>beta</sub></span>
-
-  <div class="actions">
-
-    <mat-slide-toggle color="accent" [checked]="useStorage | async" (change)="onUseStorageChanged($event)">
-      Autosave Draft
-    </mat-slide-toggle>
-
-    <button mat-icon-button [matMenuTriggerFor]="importExportMenu" matTooltip="Import/Export Data">
-      <mat-icon>import_export</mat-icon>
-    </button>
-
-    <button mat-icon-button (click)="sidebarOpen = !sidebarOpen" matTooltip="Appearance Options">
-      <mat-icon>palette</mat-icon>
-    </button>
-
-    <mat-menu #importExportMenu="matMenu">
-      <button mat-menu-item (click)="fileupload.click()">
-        <mat-icon>unarchive</mat-icon>
-        Import Data
-      </button>
-
-      <button mat-menu-item (click)="onDownload()">
-        <mat-icon>archive</mat-icon>
-        Export Data
-      </button>
-
-      <button mat-menu-item (click)="onOpenTemplates()">
-        <mat-icon>restore_page</mat-icon>
-        Use Template
-      </button>
-    </mat-menu>
-
-    <button mat-icon-button (click)="onPrint()" matTooltip="Print...">
-      <mat-icon>print</mat-icon>
-    </button>
-
-  </div>
-</mat-toolbar>
+<app-header [(sidebarOpen)]="sidebarOpen"></app-header>
 
 <mat-drawer-container class="container">
 
@@ -55,6 +16,7 @@
     <button mat-fab color="primary" class="info hidden-print" [matMenuTriggerFor]="menu">
       <mat-icon>info</mat-icon>
     </button>
+
     <mat-menu #menu="matMenu" [overlapTrigger]="true">
       <div class="info-pannel">
         <p>Made with &hearts; by <a href="https://alexfriesen.net/">Alex Friesen</a></p>
@@ -64,5 +26,3 @@
   </mat-drawer-content>
 
 </mat-drawer-container>
-
-<input class="hidden" #fileupload type="file" (change)="onUpload($event)">
