+++ src/app/app.component.scss
@@ -1,7 +1,6 @@
 :host {
   display: flex;
   flex-direction: column;
-  align-items: center;
   flex: 1;
 
   height: 100%;
@@ -51,10 +50,6 @@
   background-color: #ffffff;
 }
 
-.toolbar {
-  justify-content: space-between;
-}
-
 .info {
   position: fixed;
   bottom: 32px;
@@ -66,10 +61,6 @@
 }
 
 @media print {
-  mat-toolbar {
-    display: none;
-  }
-
   :host {
     display: block;
   }
