+++ src/app/components/header/header.component.scss
@@ -1,0 +1,3 @@
+.toolbar {
+  justify-content: space-between;
+}
