+++ src/styles.scss
@@ -10,11 +10,9 @@ body {
   font-family: Roboto, "Helvetica Neue", sans-serif;
 }
 
-
 .hidden {
   display: none;
 }
-
 
 @media print {
   .hidden-print {
